requirements.txt
streamlit
google-cloud-firestore
google-cloud-storage
google-genai
google-cloud-discoveryengine
